Changed Vector3 Type to float

+++ SphWaterfall/SphWaterfall/src/data/Vector3.cpp
@@ -5,7 +5,7 @@ Vector3::Vector3() {
 
 }
 
-Vector3::Vector3(int x, int y, int z) :
+Vector3::Vector3(float x, float y, float z) :
 	x(x),
 	y(y),
 	z(z) {
